% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timevisBasic.R
\name{timevis}
\alias{timevis}
\title{Create a basic timeline visualization}
\usage{
timevis(data = NULL, width = NULL, height = NULL, elementId = NULL)
}
\description{
Create a basic timeline visualization
}
\seealso{
\href{http://daattali.com/shiny/timevis-demo/}{Demo Shiny app}
}

